Added emulator files - complete version, no MODs
This version of the XM23 is fully functional and supports all language features covered. Devices are not yet implemented and might be added in a future update. 
This version supports the entirety of the XM23 instruction set, is fully functioning, and optimized.

Three mode cache functionality is integrated into the CPU emulator along with the full debugging suite.